+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## 0.55 (Pending):
 * Rework PersistenceTestModule constructor to pass databasePrefix as an optional parameter.
 * Remove PersistenceTestModule.registerUserTransaction() as it is unused.
-* Add abstract PersistenceTestModule.getTablesToClean() to simplify the mechanisms via which tables
-  that need cleaning are identified.
+* Pass the tables to clean as a parameter to the constructor.
 * Remove support for deriving the tables to clean in PersistenceTestModule from @Table annotations as
   domgen now explicitly lists all the tables to clean.
 * Always inject EntityListeners in PersistenceTestModule as no longer need to support EE6 and make